chore: 장소 상세 페이지 컴포넌트 분리 (#173)
